+++ src/app/not-found.tsx
@@ -1,33 +1,16 @@
-"use client";
-
+import type { Metadata } from "next";
 import Link from "next/link";
-import { useState } from "react";
+import NotFoundSubscribeForm from "@/components/NotFoundSubscribeForm";
+
+export const metadata: Metadata = {
+  title: "Page Not Found | AI Architect Series",
+  robots: {
+    index: false,
+    follow: false,
+  },
+};
 
 export default function NotFound() {
-  const [email, setEmail] = useState("");
-  const [status, setStatus] = useState<"idle" | "loading" | "success" | "error">("idle");
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!email || status === "loading") return;
-    setStatus("loading");
-    try {
-      const res = await fetch("/api/subscribe", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ email, source: "404-page" }),
-      });
-      const data = await res.json();
-      if (res.ok || data.success) {
-        setStatus("success");
-      } else {
-        setStatus("error");
-      }
-    } catch {
-      setStatus("error");
-    }
-  };
-
   return (
     <div className="min-h-screen flex flex-col items-center justify-center px-4 text-center bg-navy">
       <div className="max-w-lg w-full">
@@ -73,32 +56,7 @@ export default function NotFound() {
           <p className="text-xs text-text-secondary mb-4">
             Get weekly AI business insights delivered to your inbox.
           </p>
-          {status === "success" ? (
-            <p className="text-green-400 text-sm font-medium">Subscribed! Check your email.</p>
-          ) : (
-            <form onSubmit={handleSubmit} className="flex gap-2">
-              <label htmlFor="nf-email" className="sr-only">Email address</label>
-              <input
-                id="nf-email"
-                type="email"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                placeholder="your@email.com"
-                required
-                className="flex-1 min-w-0 rounded-xl border border-gold/20 bg-navy-dark px-3 py-2 text-sm text-text-primary placeholder:text-text-secondary/40 focus:outline-none focus:border-gold/50 transition-colors"
-              />
-              <button
-                type="submit"
-                disabled={status === "loading"}
-                className="rounded-xl bg-gold px-4 py-2 text-sm font-bold text-navy-dark hover:bg-gold-light transition-colors disabled:opacity-50 whitespace-nowrap"
-              >
-                {status === "loading" ? "..." : "Get Free"}
-              </button>
-            </form>
-          )}
-          {status === "error" && (
-            <p role="alert" className="text-red-400 text-xs mt-2">An error occurred. Please try again.</p>
-          )}
+          <NotFoundSubscribeForm />
         </div>
       </div>
     </div>
